PI Api & UI at main
I realized that relative imports is possible when calling a package from a main.py file.
With that, I was able to match my UI and the PI Api Class to the main.py file.
I can do relative move and display position in indicator.
Now I have a more formal (but simple) user interface.