LED's Working
I was having problems when trying to show up to the final user when the controller is connected and the motor is moving.
The issue was that the indicators where for some reason dependent from each other. So when one was active, the other use to shutdown.. So I thought the problem was related with Threads, but at the end (after implementing QThreads) I used to see the same problem.
So, the solution was to encapsulate one indicator in a Group box.
After that there is a little blink in the connection led, but it continues showing that is connected.
I improve how to read the moving state from the motor by reading the binary bit of the hex number received by command sent to the controller.