Added Wraps for Decorators
With this new update I added wraps functionality. So when debugging we can detect where the error is coming from.

The software works, in fact the error in threads I was having in Newport stage is fixed afters adding this update.

There is only an error at trying to Initialize the stages. That it was not shown before.